feat: exercises 4.15-23

+++ part4/bloglist_backend/controllers/blogs.js
@@ -1,19 +1,35 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
-
+const middleware = require('../utils/middleware')
 
 blogsRouter.get('/', async (request, response) => {
-  const blogs = await Blog.find({})
+  const blogs = await Blog
+    .find({})
+    .populate('user', { username: 1, name: 1, id: 1 })
   response.json(blogs)
 })
 
-blogsRouter.post('/', async (request, response) => {
-  const blog = new Blog(request.body)
+blogsRouter.post('/', middleware.userExtractor, async (request, response) => {
+  const user = request.user
+
+  const blog = new Blog({ ...request.body, user: user._id })
   const savedBlog = await blog.save()
+
+  user.blogs = user.blogs.concat(savedBlog._id)
+  await user.save()
+
   response.status(201).json(savedBlog)
 })
 
-blogsRouter.delete('/:id', async (request, response) => {
+blogsRouter.delete('/:id', middleware.userExtractor, async (request, response) => {
+  const user = request.user
+  const userMatch = user.blogs.filter(
+    blogid => blogid.toString() === request.params.id)
+
+  if (!userMatch || userMatch.length === 0) {
+    return response.status(403).json({ error: 'User does not have permission to delete this blog' })
+  }
+
   await Blog.findByIdAndRemove(request.params.id)
   response.status(204).end()
 })
